feat(notifier): reuse pending subscription
